refactor(executor): extract git and decompose functions to separate files (GH-1141)
Move 8 functions from runner.go to reduce file size:
- runner_git.go: syncMainBranch, syncNavigatorIndex, maybeInitNavigator,
  extractTaskNumber, containsTaskNumber, ExtractRepoName, ValidateRepoProjectMatch
- runner_decompose.go: executeDecomposedTask

This reduces runner.go from ~3,370 lines to ~3,167 lines.
