+++ internal/executor/runner.go
@@ -2202,209 +2202,6 @@ The previous execution completed but made no code changes. This task requires ac
 
 	return result, nil
 }
-
-// executeDecomposedTask runs subtasks sequentially and aggregates results (GH-218).
-// Each subtask runs to completion before the next starts. Only the final subtask
-// creates a PR (CreatePR is already set by the decomposer). All changes accumulate
-// on the same branch, so the final PR contains all subtask work.
-func (r *Runner) executeDecomposedTask(ctx context.Context, parentTask *Task, subtasks []*Task) (*ExecutionResult, error) {
-	start := time.Now()
-	totalSubtasks := len(subtasks)
-
-	r.log.Info("Starting decomposed task execution",
-		slog.String("parent_id", parentTask.ID),
-		slog.Int("subtask_count", totalSubtasks),
-	)
-
-	// Emit parent task started event
-	r.emitAlertEvent(AlertEvent{
-		Type:      AlertEventTypeTaskStarted,
-		TaskID:    parentTask.ID,
-		TaskTitle: parentTask.Title,
-		Project:   parentTask.ProjectPath,
-		Metadata: map[string]string{
-			"decomposed":    "true",
-			"subtask_count": fmt.Sprintf("%d", totalSubtasks),
-		},
-		Timestamp: time.Now(),
-	})
-
-	// Dispatch webhook for decomposed task started
-	r.dispatchWebhook(ctx, webhooks.EventTaskStarted, webhooks.TaskStartedData{
-		TaskID:      parentTask.ID,
-		Title:       parentTask.Title,
-		Description: fmt.Sprintf("Decomposed into %d subtasks: %s", totalSubtasks, parentTask.Description),
-		Project:     parentTask.ProjectPath,
-		Source:      "pilot",
-	})
-
-	// Initialize git and create branch ONCE for all subtasks
-	git := NewGitOperations(parentTask.ProjectPath)
-	if parentTask.Branch != "" {
-		r.reportProgress(parentTask.ID, "Branching", 1, "Switching to default branch...")
-
-		// GH-279: Always switch to default branch and pull latest before creating new branch.
-		// This prevents new branches from forking off previous pilot branches instead of main.
-		// GH-836: Hard fail if we can't switch - continuing from wrong branch causes corrupted PRs.
-		defaultBranch, err := git.SwitchToDefaultBranchAndPull(ctx)
-		if err != nil {
-			return nil, fmt.Errorf("branch switch failed, aborting execution: failed to switch to default branch: %w", err)
-		}
-		r.reportProgress(parentTask.ID, "Branching", 2, fmt.Sprintf("On %s, creating %s...", defaultBranch, parentTask.Branch))
-
-		if err := git.CreateBranch(ctx, parentTask.Branch); err != nil {
-			if switchErr := git.SwitchBranch(ctx, parentTask.Branch); switchErr != nil {
-				return nil, fmt.Errorf("failed to create/switch branch: %w", err)
-			}
-		}
-		r.reportProgress(parentTask.ID, "Branching", 5, fmt.Sprintf("Branch %s ready", parentTask.Branch))
-	}
-
-	// Aggregate result
-	aggregateResult := &ExecutionResult{
-		TaskID:  parentTask.ID,
-		Success: true,
-	}
-
-	// Execute each subtask sequentially
-	for i, subtask := range subtasks {
-		subtaskNum := i + 1
-
-		// Report progress with subtask counter
-		progressPct := 5 + (85 * subtaskNum / totalSubtasks)
-		r.reportProgress(parentTask.ID, "Decomposed", progressPct,
-			fmt.Sprintf("Subtask %d/%d: %s", subtaskNum, totalSubtasks, truncateText(subtask.Title, 40)))
-
-		r.log.Info("Executing subtask",
-			slog.String("parent_id", parentTask.ID),
-			slog.String("subtask_id", subtask.ID),
-			slog.Int("index", subtaskNum),
-			slog.Int("total", totalSubtasks),
-		)
-
-		// Execute subtask (recursively calls Execute, but subtasks won't decompose further)
-		// Clear the branch since we already created it
-		subtask.Branch = ""
-
-		// Temporarily disable decomposer to prevent recursive decomposition
-		savedDecomposer := r.decomposer
-		r.decomposer = nil
-
-		subtaskResult, err := r.executeWithOptions(ctx, subtask, false)
-
-		// Restore decomposer
-		r.decomposer = savedDecomposer
-
-		if err != nil {
-			r.log.Error("Subtask execution error",
-				slog.String("subtask_id", subtask.ID),
-				slog.Any("error", err),
-			)
-			aggregateResult.Success = false
-			aggregateResult.Error = fmt.Sprintf("subtask %d/%d failed: %v", subtaskNum, totalSubtasks, err)
-			break
-		}
-
-		if !subtaskResult.Success {
-			r.log.Warn("Subtask failed",
-				slog.String("subtask_id", subtask.ID),
-				slog.String("error", subtaskResult.Error),
-			)
-			aggregateResult.Success = false
-			aggregateResult.Error = fmt.Sprintf("subtask %d/%d failed: %s", subtaskNum, totalSubtasks, subtaskResult.Error)
-			break
-		}
-
-		// Aggregate metrics
-		aggregateResult.TokensInput += subtaskResult.TokensInput
-		aggregateResult.TokensOutput += subtaskResult.TokensOutput
-		aggregateResult.TokensTotal += subtaskResult.TokensTotal
-		aggregateResult.ResearchTokens += subtaskResult.ResearchTokens
-		aggregateResult.FilesChanged += subtaskResult.FilesChanged
-		aggregateResult.LinesAdded += subtaskResult.LinesAdded
-		aggregateResult.LinesRemoved += subtaskResult.LinesRemoved
-
-		// Keep last commit SHA and PR URL
-		if subtaskResult.CommitSHA != "" {
-			aggregateResult.CommitSHA = subtaskResult.CommitSHA
-		}
-		if subtaskResult.PRUrl != "" {
-			aggregateResult.PRUrl = subtaskResult.PRUrl
-		}
-		if subtaskResult.ModelName != "" {
-			aggregateResult.ModelName = subtaskResult.ModelName
-		}
-
-		// Track quality gates from final subtask
-		if subtask.CreatePR && subtaskResult.QualityGates != nil {
-			aggregateResult.QualityGates = subtaskResult.QualityGates
-		}
-
-		r.log.Info("Subtask completed",
-			slog.String("subtask_id", subtask.ID),
-			slog.Int("index", subtaskNum),
-			slog.Int("total", totalSubtasks),
-		)
-	}
-
-	aggregateResult.Duration = time.Since(start)
-	aggregateResult.EstimatedCostUSD = estimateCost(
-		aggregateResult.TokensInput+aggregateResult.ResearchTokens,
-		aggregateResult.TokensOutput,
-		aggregateResult.ModelName,
-	)
-
-	// Emit completion event
-	if aggregateResult.Success {
-		r.reportProgress(parentTask.ID, "Completed", 100,
-			fmt.Sprintf("All %d subtasks completed", totalSubtasks))
-
-		r.emitAlertEvent(AlertEvent{
-			Type:      AlertEventTypeTaskCompleted,
-			TaskID:    parentTask.ID,
-			TaskTitle: parentTask.Title,
-			Project:   parentTask.ProjectPath,
-			Metadata: map[string]string{
-				"duration_ms":   fmt.Sprintf("%d", aggregateResult.Duration.Milliseconds()),
-				"pr_url":        aggregateResult.PRUrl,
-				"subtask_count": fmt.Sprintf("%d", totalSubtasks),
-			},
-			Timestamp: time.Now(),
-		})
-
-		r.dispatchWebhook(ctx, webhooks.EventTaskCompleted, webhooks.TaskCompletedData{
-			TaskID:    parentTask.ID,
-			Title:     parentTask.Title,
-			Project:   parentTask.ProjectPath,
-			Duration:  aggregateResult.Duration,
-			PRCreated: aggregateResult.PRUrl != "",
-			PRURL:     aggregateResult.PRUrl,
-		})
-	} else {
-		r.reportProgress(parentTask.ID, "Failed", 100, aggregateResult.Error)
-
-		r.emitAlertEvent(AlertEvent{
-			Type:      AlertEventTypeTaskFailed,
-			TaskID:    parentTask.ID,
-			TaskTitle: parentTask.Title,
-			Project:   parentTask.ProjectPath,
-			Error:     aggregateResult.Error,
-			Timestamp: time.Now(),
-		})
-
-		r.dispatchWebhook(ctx, webhooks.EventTaskFailed, webhooks.TaskFailedData{
-			TaskID:   parentTask.ID,
-			Title:    parentTask.Title,
-			Project:  parentTask.ProjectPath,
-			Duration: aggregateResult.Duration,
-			Error:    aggregateResult.Error,
-			Phase:    "Decomposed",
-		})
-	}
-
-	return aggregateResult, nil
-}
-
 // Cancel terminates a running task by killing its Claude Code process.
 // Returns an error if the task is not currently running.
 func (r *Runner) Cancel(taskID string) error {
@@ -3305,219 +3102,6 @@ func (r *Runner) reportProgress(taskID, phase string, progress int, message stri
 	}
 }
 
-// maybeInitNavigator initializes Navigator structure if not present.
-// This enables auto-init for projects that don't have .agent/ yet.
-func (r *Runner) maybeInitNavigator(projectPath string) error {
-	initializer, err := NewNavigatorInitializer(r.log)
-	if err != nil {
-		return fmt.Errorf("failed to create navigator initializer: %w", err)
-	}
-
-	if initializer.IsInitialized(projectPath) {
-		return nil
-	}
-
-	r.log.Info("Auto-initializing Navigator for project",
-		slog.String("path", projectPath),
-	)
-
-	return initializer.Initialize(projectPath)
-}
-
-// syncNavigatorIndex updates DEVELOPMENT-README.md after task completion.
-// It moves completed tasks from "In Progress" to "Completed" section.
-// Supports both TASK-XX and GH-XX formats.
-func (r *Runner) syncNavigatorIndex(task *Task, status string, executionPath string) error {
-	indexPath := filepath.Join(executionPath, ".agent", "DEVELOPMENT-README.md")
-
-	// Check if Navigator index exists
-	if _, err := os.Stat(indexPath); os.IsNotExist(err) {
-		r.log.Debug("Navigator index not found, skipping sync",
-			slog.String("task_id", task.ID),
-			slog.String("path", indexPath),
-		)
-		return nil
-	}
-
-	// Read current index
-	content, err := os.ReadFile(indexPath)
-	if err != nil {
-		return fmt.Errorf("failed to read navigator index: %w", err)
-	}
-
-	lines := strings.Split(string(content), "\n")
-	var result []string
-	var taskEntry string
-	var taskTitle string
-	inProgressSection := false
-	completedSection := false
-	completedInsertIdx := -1
-
-	// Extract task number for matching (handles GH-XX and TASK-XX)
-	taskNum := extractTaskNumber(task.ID)
-
-	for i, line := range lines {
-		// Track sections
-		if strings.Contains(line, "### In Progress") {
-			inProgressSection = true
-			completedSection = false
-			result = append(result, line)
-			continue
-		}
-		if strings.Contains(line, "### Backlog") || strings.Contains(line, "## Completed") {
-			inProgressSection = false
-		}
-		if strings.Contains(line, "## Completed") {
-			completedSection = true
-			result = append(result, line)
-			// Find where to insert (after the header and any date line)
-			completedInsertIdx = len(result)
-			continue
-		}
-		if completedSection && strings.HasPrefix(strings.TrimSpace(line), "| Item") {
-			// After table header in completed section
-			result = append(result, line)
-			completedInsertIdx = len(result)
-			continue
-		}
-		if completedSection && strings.HasPrefix(strings.TrimSpace(line), "|---") {
-			result = append(result, line)
-			completedInsertIdx = len(result)
-			continue
-		}
-
-		// Check if this line contains our task in the In Progress table
-		if inProgressSection && strings.Contains(line, "|") {
-			// Table row format: | GH# | Title | Status |
-			// or: | 54 | Speed Optimization ... | 🔄 Pilot executing |
-			if strings.Contains(line, task.ID) || (taskNum != "" && containsTaskNumber(line, taskNum)) {
-				// Extract title from the row
-				parts := strings.Split(line, "|")
-				if len(parts) >= 3 {
-					taskTitle = strings.TrimSpace(parts[2]) // Title is second column after GH#
-				}
-				taskEntry = line
-				// Skip this line (don't add to result) - we'll move it to completed
-				continue
-			}
-		}
-
-		result = append(result, line)
-		_ = i // suppress unused warning
-	}
-
-	// If we found a task entry to move
-	if taskEntry != "" && completedInsertIdx >= 0 {
-		// Create completed entry
-		completedEntry := fmt.Sprintf("| %s | %s |", task.ID, taskTitle)
-
-		// Insert at the right position
-		newResult := make([]string, 0, len(result)+1)
-		newResult = append(newResult, result[:completedInsertIdx]...)
-		newResult = append(newResult, completedEntry)
-		newResult = append(newResult, result[completedInsertIdx:]...)
-		result = newResult
-
-		// Write updated index
-		if err := os.WriteFile(indexPath, []byte(strings.Join(result, "\n")), 0644); err != nil {
-			return fmt.Errorf("failed to write navigator index: %w", err)
-		}
-
-		r.log.Info("Updated Navigator index",
-			slog.String("task_id", task.ID),
-			slog.String("status", status),
-			slog.String("moved_to", "Completed"),
-		)
-	} else if taskEntry != "" {
-		r.log.Debug("Task found but no Completed section to move to",
-			slog.String("task_id", task.ID),
-		)
-	} else {
-		r.log.Debug("Task not found in Navigator index In Progress section",
-			slog.String("task_id", task.ID),
-		)
-	}
-
-	return nil
-}
-
-// syncMainBranch syncs the local main branch with origin/main after task completion.
-// GH-1018: This prevents local/remote divergence over time when multiple PRs are merged.
-//
-// Strategy:
-// 1. Fetch origin/main to get latest remote state
-// 2. If on main branch, reset --hard to origin/main
-// 3. If on feature branch, skip reset (don't disrupt worktree)
-//
-// This is opt-in via executor.sync_main_after_task config.
-func (r *Runner) syncMainBranch(ctx context.Context, repoPath string) error {
-	log := r.log.With(slog.String("repo", repoPath))
-	log.Debug("Syncing main branch with origin")
-
-	// Fetch latest from origin
-	fetchCmd := exec.CommandContext(ctx, "git", "fetch", "origin", "main")
-	fetchCmd.Dir = repoPath
-	if output, err := fetchCmd.CombinedOutput(); err != nil {
-		return fmt.Errorf("failed to fetch origin/main: %w: %s", err, output)
-	}
-
-	// Check current branch
-	branchCmd := exec.CommandContext(ctx, "git", "rev-parse", "--abbrev-ref", "HEAD")
-	branchCmd.Dir = repoPath
-	branchOutput, err := branchCmd.Output()
-	if err != nil {
-		return fmt.Errorf("failed to get current branch: %w", err)
-	}
-
-	currentBranch := strings.TrimSpace(string(branchOutput))
-
-	// Only reset if on main branch (don't disrupt feature branches or worktrees)
-	if currentBranch == "main" || currentBranch == "master" {
-		resetCmd := exec.CommandContext(ctx, "git", "reset", "--hard", "origin/main")
-		resetCmd.Dir = repoPath
-		if output, err := resetCmd.CombinedOutput(); err != nil {
-			return fmt.Errorf("failed to reset main to origin/main: %w: %s", err, output)
-		}
-		log.Info("Synced main branch with origin/main")
-	} else {
-		log.Debug("Not on main branch, skipping reset", slog.String("branch", currentBranch))
-	}
-
-	return nil
-}
-
-// extractTaskNumber extracts the numeric part from task IDs like "GH-57" or "TASK-123"
-func extractTaskNumber(taskID string) string {
-	// Handle GH-XX format
-	if strings.HasPrefix(taskID, "GH-") {
-		return strings.TrimPrefix(taskID, "GH-")
-	}
-	// Handle TASK-XX format
-	if strings.HasPrefix(taskID, "TASK-") {
-		return strings.TrimPrefix(taskID, "TASK-")
-	}
-	return taskID
-}
-
-// containsTaskNumber checks if a line contains a task number in various formats
-func containsTaskNumber(line, taskNum string) bool {
-	// Check for "| 57 |" or "| GH-57 |" or "| TASK-57 |" patterns
-	patterns := []string{
-		fmt.Sprintf("| %s ", taskNum),
-		fmt.Sprintf("|%s ", taskNum),
-		fmt.Sprintf("| %s|", taskNum),
-		fmt.Sprintf("|%s|", taskNum),
-		fmt.Sprintf("GH-%s", taskNum),
-		fmt.Sprintf("TASK-%s", taskNum),
-	}
-	for _, p := range patterns {
-		if strings.Contains(line, p) {
-			return true
-		}
-	}
-	return false
-}
-
 // buildQualityGatesResult converts QualityOutcome to QualityGatesResult for ExecutionResult (GH-209)
 func (r *Runner) buildQualityGatesResult(outcome *QualityOutcome, totalRetries int) *QualityGatesResult {
 	if outcome == nil {
@@ -3580,39 +3164,4 @@ func (c *simpleQualityChecker) Check(ctx context.Context) (*QualityOutcome, erro
 	}
 
 	return outcome, nil
-}
-
-// ExtractRepoName extracts the repository name from "owner/repo" format.
-// Returns just the repo part (e.g., "pilot" from "alekspetrov/pilot").
-func ExtractRepoName(repo string) string {
-	parts := strings.Split(repo, "/")
-	if len(parts) == 2 {
-		return parts[1]
-	}
-	return repo
-}
-
-// ValidateRepoProjectMatch validates that a source repo matches the project path.
-// This is a defense against cross-project execution (GH-386).
-// Returns an error if there's a mismatch between repo name and project directory name.
-func ValidateRepoProjectMatch(sourceRepo, projectPath string) error {
-	if sourceRepo == "" || projectPath == "" {
-		return nil // Nothing to validate
-	}
-
-	repoName := ExtractRepoName(sourceRepo)
-	projectDir := filepath.Base(projectPath)
-
-	// Normalize for comparison (case-insensitive)
-	repoName = strings.ToLower(repoName)
-	projectDir = strings.ToLower(projectDir)
-
-	if repoName != projectDir {
-		return fmt.Errorf(
-			"repo/project mismatch: issue from '%s' but executing in '%s' (expected project directory '%s')",
-			sourceRepo, projectPath, repoName,
-		)
-	}
-
-	return nil
 }
